requirements.txt
flask
python-dotenv
azure-ai-formrecognizer
azure-core
azure-cosmos